[feat] api 비동기 처리
