+++ src/main/java/com/demoday/ddangddangddang/service/third/JudgeFacade.java
@@ -19,52 +19,17 @@ import org.springframework.stereotype.Service;
 public class JudgeFacade {
 
     private final FinalJudgeService finalJudgeService;
-    private final ChatGptService chatGptService;
-    private final VoteRepository voteRepository;
-    private final SseEmitters sseEmitters;
+    private final JudgmentAsyncExecutor judgmentAsyncExecutor; // ✨ 새로 주입
 
-    // 1. 동기: 클라이언트 요청 접수 및 비동기 처리 시작
     public ApiResponse<String> requestFinalJudge(Long caseId, FinalJudgmentRequestDto voteDto, Long userId) {
-        // Read-only 트랜잭션으로 데이터 준비
+        // 1. Read-only 트랜잭션으로 데이터 준비
         JudgeContextDto context = finalJudgeService.prepareJudgeContext(caseId, userId);
 
-        // 비동기 작업 시작
-        processAsyncJudgment(caseId, voteDto, context);
+        // 2. 비동기 작업 시작 (여기서 더 이상 안 기다림)
+        judgmentAsyncExecutor.processAsyncJudgment(caseId, voteDto, context);
 
+        // 3. 클라이언트에게는 바로 응답
         return ApiResponse.onSuccess("판결 요청이 접수되었습니다.", "JUDGMENT_IN_PROGRESS");
     }
 
-    // 2. 비동기: AI 호출 -> DB 저장 -> SSE 알림
-    @Async
-    public void processAsyncJudgment(Long caseId, FinalJudgmentRequestDto voteDto, JudgeContextDto context) {
-        long startTime = System.currentTimeMillis();
-        log.info("[Async] 판결 프로세스 시작: caseId={}", caseId);
-        try {
-            // A. AI 호출 (병목 구간)
-            long apiCallStart = System.currentTimeMillis();
-            AiJudgmentDto aiResult = chatGptService.requestFinalJudgment(context.getACase(),
-                    context.getAdoptedDefenses(),
-                    context.getAdoptedRebuttals(),
-                    voteDto.getVotesA(),
-                    voteDto.getVotesB());
-            long apiCallEnd = System.currentTimeMillis();
-            log.warn("⏱️ [측정] ChatGPT API 호출 시간: {}ms", apiCallEnd - apiCallStart); // API 호출 시간 측정
-
-            // B. DB 저장 및 정산
-            long dbStart = System.currentTimeMillis();
-            Long judgmentId = finalJudgeService.saveJudgeResultAndSettle(caseId, aiResult);
-            long dbEnd = System.currentTimeMillis();
-            log.warn("⏱️ [측정] DB 저장 및 정산 시간: {}ms", dbEnd - dbStart); // DB 정산 시간 측정
-
-            // C. SSE 알림 전송
-            sseEmitters.send(caseId, judgmentId);
-
-            long endTime = System.currentTimeMillis();
-            log.info("[Async] 총 비동기 작업 시간: {}ms", endTime - startTime); // 총 비동기 작업 시간
-
-        } catch (Exception e) {
-            log.error("[Async] 판결 프로세스 실패: caseId={}, error={}", caseId, e.getMessage(), e);
-            sseEmitters.sendError(caseId, "판결 생성 중 오류가 발생했습니다.");
-        }
-    }
 }
